Separate memory from CPU object
Creates a memory object in main.py, then passes that object by reference to the CPU, so that the other components all have access to the same memory. main.py also now reads the ROM into memory.

This commit also moves the utility functions like toUnsignedDecimal() to functions.py so that they can be used by all of the systems components.